backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
python-multipart==0.0.6
pyannote.audio==3.1.0
faster-whisper==1.0.0
librosa==0.10.1
pydub==0.25.1
openai==1.40.3
structlog==23.1.0
torch==2.0.1
torchaudio==2.0.2
numpy==1.24.3
pydantic-settings==2.10.1
python-dotenv==1.1.1
speechbrain==0.5.16